Delete old pass
