+++ cpg-concepts/src/integrationTest/kotlin/de/fraunhofer/aisec/cpg/concepts/file/FileConceptTest.kt
@@ -28,7 +28,7 @@ package de.fraunhofer.aisec.cpg.concepts.file
 import de.fraunhofer.aisec.cpg.frontends.python.PythonLanguage
 import de.fraunhofer.aisec.cpg.graph.*
 import de.fraunhofer.aisec.cpg.graph.concepts.file.*
-import de.fraunhofer.aisec.cpg.passes.concepts.file.python.PythonFileEOGConceptPass
+import de.fraunhofer.aisec.cpg.passes.concepts.file.python.PythonFileConceptPass
 import de.fraunhofer.aisec.cpg.query.Must
 import de.fraunhofer.aisec.cpg.query.dataFlow
 import de.fraunhofer.aisec.cpg.query.executionPath
@@ -53,7 +53,7 @@ class FileConceptTest : BaseTest() {
                 usePasses = true,
             ) {
                 it.registerLanguage<PythonLanguage>()
-                it.registerPass<PythonFileEOGConceptPass>()
+                it.registerPass<PythonFileConceptPass>()
                 it.symbols(mapOf("PYTHON_PLATFORM" to "linux"))
             }
         assertNotNull(result)
@@ -133,7 +133,7 @@ class FileConceptTest : BaseTest() {
                 usePasses = true,
             ) {
                 it.registerLanguage<PythonLanguage>()
-                it.registerPass<PythonFileEOGConceptPass>()
+                it.registerPass<PythonFileConceptPass>()
                 it.symbols(mapOf("PYTHON_PLATFORM" to "linux"))
             }
         assertNotNull(result)
@@ -186,7 +186,7 @@ class FileConceptTest : BaseTest() {
                 usePasses = true,
             ) {
                 it.registerLanguage<PythonLanguage>()
-                it.registerPass<PythonFileEOGConceptPass>()
+                it.registerPass<PythonFileConceptPass>()
                 it.symbols(mapOf("PYTHON_PLATFORM" to "linux"))
             }
         assertNotNull(result)
@@ -240,7 +240,7 @@ class FileConceptTest : BaseTest() {
                 usePasses = true,
             ) {
                 it.registerLanguage<PythonLanguage>()
-                it.registerPass<PythonFileEOGConceptPass>()
+                it.registerPass<PythonFileConceptPass>()
                 it.symbols(mapOf("PYTHON_PLATFORM" to "linux"))
             }
         assertNotNull(result)
@@ -288,7 +288,7 @@ class FileConceptTest : BaseTest() {
                 usePasses = true,
             ) {
                 it.registerLanguage<PythonLanguage>()
-                it.registerPass<PythonFileEOGConceptPass>()
+                it.registerPass<PythonFileConceptPass>()
                 it.symbols(mapOf("PYTHON_PLATFORM" to "linux"))
             }
         assertNotNull(result)
@@ -329,7 +329,7 @@ class FileConceptTest : BaseTest() {
                 usePasses = true,
             ) {
                 it.registerLanguage<PythonLanguage>()
-                it.registerPass<PythonFileEOGConceptPass>()
+                it.registerPass<PythonFileConceptPass>()
                 it.symbols(mapOf("PYTHON_PLATFORM" to "linux"))
             }
         assertNotNull(result)
@@ -359,7 +359,7 @@ class FileConceptTest : BaseTest() {
                 usePasses = true,
             ) {
                 it.registerLanguage<PythonLanguage>()
-                it.registerPass<PythonFileEOGConceptPass>()
+                it.registerPass<PythonFileConceptPass>()
                 it.symbols(mapOf("PYTHON_PLATFORM" to "linux"))
             }
         assertNotNull(result)
